pyproject.toml
[project]
name = "cf-yx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cloudflare>=5.1.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
